backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
drf-spectacular>=0.27.0
django-filter>=23.0
Pillow>=10.0.0
python-decouple>=3.8
django-cors-headers>=4.3.0

# Testing packages
pytest>=7.4.0
pytest-django>=4.5.0
pytest-xdist>=3.3.0  # For parallel test execution
pytest-cov>=4.1.0  # For coverage reports
locust>=2.17.0  # For load/stress testing
factory-boy>=3.3.0  # For test data factories
faker>=20.0.0  # For generating fake data
freezegun>=1.2.0  # For mocking datetime in tests

# Code quality and linting
flake8>=6.0.0  # For linting
black>=23.0.0  # For code formatting
isort>=5.12.0  # For import sorting
bandit>=1.7.5  # For security linting
safety>=2.3.0  # For dependency vulnerability checking
